% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalegm.R
\name{install_causalegm}
\alias{install_causalegm}
\title{Install the python CausalEGM package}
\usage{
install_causalegm(method = "auto", pip = TRUE)
}
\arguments{
\item{method}{default "auto"}

\item{pip}{boolean flag, default TRUE}
}
\value{
No return value
}
\description{
Install the python CausalEGM package
}
